Rename pycarwings2 to pycarwings3
