--- pycarwings2/__init__.py
@@ -1,4 +1,0 @@
-__all__ = ['pycarwings2', 'responses']
-from .pycarwings2 import Session, Leaf, CarwingsError  # noqa: F401
-
-__version__ = '0.0.1'
