+++ setup.cfg
@@ -2,7 +2,7 @@
 test=pytest
 
 [tool:pytest]
-addopts = --flake8 --cov=pycarwings2 --cache-clear --ignore=venv --verbose
+addopts = --flake8 --cov=pycarwings3 --cache-clear --ignore=venv --verbose
 flake8-ignore = 
 	bin ALL
 	lib ALL
